Move internal/director to pkg/director in System Broker
