+++ components/system-broker/internal/osb/bind_delete.go
@@ -21,7 +21,7 @@ import (
 
 	"github.com/kyma-incubator/compass/components/system-broker/pkg/types"
 
-	"github.com/kyma-incubator/compass/components/system-broker/internal/director"
+	"github.com/kyma-incubator/compass/components/system-broker/pkg/director"
 	"github.com/pkg/errors"
 
 	"github.com/kyma-incubator/compass/components/director/pkg/log"
